Adds a popup which can trigger an arbitrary function.
As a result of the correct function binding, we can build our own HTML
form to go inside a popup.

This HTML form of course can have custom styling.

This HTML form can then trigger whatever events we want - for instance,
it could take in a message and then update the Vuex store.